add details to documentation & example
