+++ docs/src/examples/6_spectrahedron.jl
